WIP - added SessionDaoTest
