+++ data/src/test/java/com/android254/data/repos/SampleData.kt
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2022 DroidconKE
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.android254.data.repos
+
+import com.android254.data.network.models.responses.PaginatedResponse
+import com.android254.data.network.models.responses.PaginationMetaData
+import com.android254.data.network.models.responses.ResponseMetaData
+import com.android254.data.network.models.responses.SessionApiModel
+
+val results = PaginatedResponse(
+    data = listOf(
+        SessionApiModel(
+            "1",
+            "",
+            "",
+            "",
+            "",
+            speakers = listOf(),
+            "new title"
+        )
+    ),
+    meta = ResponseMetaData(
+        paginator = PaginationMetaData(
+            1,
+            1,
+            false,
+            "",
+            "",
+            "200",
+            ""
+        )
+    )
+)
